+++ internal/adapter.go
@@ -1,6 +1,8 @@
 package internal
 
 type Adapter interface {
+	TableName() string
+	RowName() string
 	Init(url string)
 	FetchTables() (tables []table)
 	FetchTableData(table table, limit int) ([]string, [][]string)