+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added support for SQL Server
 - Added support for MongoDB
 - Added support for Redis
+- Added support for Elasticsearch
 
 ## 0.1.3 (2022-09-04)