+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.1.4 (unreleased)
+## 0.1.4 (2022-09-07)
 
 - Added support for SQL Server
 - Added support for MongoDB